fixed dd_weather_predict_witfunctions_win

+++ 20160126_DNN_Weather_Predict_withfunctions-win.py
@@ -24,48 +24,67 @@ def create_h2o_urd_model(urd_data):
     h2o.init(strict_version_check=False)
     h2o.remove_all()
 
-    # Set start and end dates for training
-    date_start = '2006-Jan-01'
-    date_end = '2014-Dec-31'
-
-    # Find row indices of training data
-    start_row = urd_data[urd_data['Date1'] == date_start].index.tolist()[0]
-    end_row = urd_data[urd_data['Date1'] == date_end].index.tolist()[0]
-
-    # Create training data slice and convert to training and validation H2OFrames
-    urd_data_pd_train = urd_data[start_row:end_row + 1].copy()
-    urd_data_pd_train_nodate = urd_data_pd_train.drop('Date1', axis=1, inplace=False)
-    train = h2o.H2OFrame(urd_data_pd_train_nodate,
-                         column_types=['int', 'enum', 'real', 'real', 'int', 'int', 'int', 'int'])
-    training, validation = train.split_frame(ratios=[0.8])
-
-    # Define predictors and response
-    predictors = list(train.columns)[2:]
-    response = list(train.columns)[0]
-
-    # Run DNN
+    # Define path to model
     urd_model_id = 'Python_URD_DNN_2006-2014'
-    urd_model = H2ODeepLearningEstimator(model_id=urd_model_id, epochs=5000, hidden=[800, 800], activation="Tanh",
-                                     l1=0, l2=0, stopping_rounds=5, stopping_metric='MSE', stopping_tolerance=1e-6)
-    urd_model.train(x=predictors, y=response, training_frame=training, validation_frame=validation)
+    save_path = os.path.join('C:\\', '0MyDataBases/7R/ADHOC_Qlikview-linux/H2O_Models/')
 
-    # Save DNN model
-    save_path = 'C:\\0MyDataBases\\7R\ADHOC_Qlikview-linux\H2O_Models\\'
-    try:
-        h2o.save_model(urd_model, path=save_path)
-    except exceptions.H2OServerError:  # Raised if file already exists
-        os.remove(save_path + urd_model_id)
-        h2o.save_model(urd_model, path=save_path)
+    # Check if model exists and prompt for overwrite if exists
+    skip_h2o = None
+    if os.path.isfile(os.path.join(save_path, urd_model_id)):
+        while skip_h2o != "Y" and skip_h2o != "n":
+            skip_h2o = raw_input("H2O model {} already exists. Use existing model?(Y/n): ".format(urd_model_id))
+    else:
+        skip_h2o = 'n'
+
+    if skip_h2o == 'Y':
+
+        # Load model
+        urd_model = h2o.load_model(os.path.join(save_path, urd_model_id))
+
+    elif skip_h2o == 'n':
+
+        # Set start and end dates for training
+        date_start = '2006-Jan-01'
+        date_end = '2014-Dec-31'
+
+        # Find row indices of training data
+        start_row = urd_data[urd_data['Date1'] == date_start].index.tolist()[0]
+        end_row = urd_data[urd_data['Date1'] == date_end].index.tolist()[0]
+
+        # Create training data slice and convert to training and validation H2OFrames
+        urd_data_pd_train = urd_data[start_row:end_row + 1].copy()
+        urd_data_pd_train_nodate = urd_data_pd_train.drop('Date1', axis=1, inplace=False)
+        train = h2o.H2OFrame(urd_data_pd_train_nodate,
+                             column_types=['int', 'enum', 'real', 'real', 'int', 'int', 'int', 'int'],
+                             destination_frame='Training_Validation_Frame')
+        training, validation = train.split_frame(ratios=[0.8])
+
+        # Define predictors and response
+        predictors = list(train.columns)[2:]
+        response = list(train.columns)[0]
+
+        # Run DNN
+        urd_model = H2ODeepLearningEstimator(model_id=urd_model_id, epochs=5000, hidden=[800, 800], activation="Tanh",
+                                         l1=0, l2=0, stopping_rounds=5, stopping_metric='MSE', stopping_tolerance=1e-6)
+        urd_model.train(x=predictors, y=response, training_frame=training, validation_frame=validation)
+
+        # Save DNN model
+        try:
+            h2o.save_model(urd_model, path=save_path)
+        except exceptions.H2OServerError:  # Raised if file already exists
+            os.remove(save_path + urd_model_id)
+            h2o.save_model(urd_model, path=save_path)
 
     return urd_model
 
 
-def get_predictions(urd_model, l_test_data):
+def get_predictions(urd_model, l_test_data, l_test_names):
     """Creates predictions on test data using H2O model
 
     Args:
         urd_model (H2ODeepLearningEstimator): H2O model for prediction.
         l_test_data (list): list of pandas DataFrames to predict on.
+        l_test_names (list): list of csv files corresponding to the DataFrames
 
     Returns:
         list: List of pandas DataFrames containing predictions.
@@ -76,10 +95,12 @@ def get_predictions(urd_model, l_test_data):
     predictions = []
 
     # Convert test data to H2OFrames and run predictions
-    for test_data in l_test_data:
+    for i in range(len(l_test_data)):
+        test_data = l_test_data[i]
         test_data_no_date = test_data.drop('Date1', axis=1, inplace=False)
         h2o_test = h2o.H2OFrame(test_data_no_date,
-                                column_types=['int', 'enum', 'real', 'real', 'int', 'int', 'int', 'int'])
+                                column_types=['int', 'enum', 'real', 'real', 'int', 'int', 'int', 'int'],
+                                destination_frame=l_test_names[i] + "_Prediction")
         predictions.append(urd_model.predict(h2o_test).as_data_frame())
 
     return predictions
@@ -127,17 +148,17 @@ def aggregate_by_day_month_year(dataframe):
 if __name__ == "__main__":
 
     # Define home directory
-    home_path = os.environ.get("HOME")
+    home_path = os.path.expanduser("~")
 
     # Import URD data
-    urd_path = 'C:\\0MyDataBases\\7R\ADHOC_Qlikview-linux\data_2015\ExportFileR.csv'
+    urd_path = os.path.join('C:\\', '0MyDataBases/7R/ADHOC_Qlikview-linux/data_2015/ExportFileR.csv')
     data_full = pd.read_csv(urd_path)
 
     # Create H2O model
     model = create_h2o_urd_model(data_full)
 
     # Define list of pandas DataFrames for model to predict on
-    base_data_path = 'C:\\0MyDataBases\\7R\ADHOC_Qlikview-linux\data_2015'
+    base_data_path = os.path.join('C:\\', '0MyDataBases/7R/ADHOC_Qlikview-linux/data_2015')
     l_csv_test_data = ['ExportFileWeather_2015.csv', 'ExportFileWeather_2014.csv', 'ExportFileWeather_2013.csv',
                        'ExportFileWeather_2012.csv', 'ExportFileWeather_2011.csv', 'ExportFileWeather_2010.csv']
     l_pd_test_data = [data_full]
@@ -145,7 +166,7 @@ if __name__ == "__main__":
         l_pd_test_data.append(pd.read_csv(os.path.join(base_data_path, csv_test_data)))
 
     # Get model predictions on test data
-    l_predictions_raw = get_predictions(model, l_pd_test_data)
+    l_predictions_raw = get_predictions(model, l_pd_test_data, ['ExportFileR.csv'] + l_csv_test_data)
 
     # Add prediction column to existing pandas DataFrames
     for i in range(len(l_predictions_raw)):
@@ -157,7 +178,7 @@ if __name__ == "__main__":
         l_predictions.append(aggregate_by_day_month_year(pd_test_data))
 
     # Create list of strings indicating year of test data being used
-    l_test_year = ['2016']
+    l_test_year = ['Actual']
     for filename in l_csv_test_data:
         l_test_year.append(filename[-8:-4])
 
@@ -175,15 +196,16 @@ if __name__ == "__main__":
                 d_traces[weather_year][time_frame]['Prediction'] = go.Scatter(
                     x=d_predictions[weather_year][time_frame]['Date'],
                     y=d_predictions[weather_year][time_frame]['Prediction'],
-                    name=time_frame + ' Prediction ' + weather_year,
+                    name="{0} Prediction ({1} Weather)".format(time_frame, weather_year),
                     line=dict(dash='dash'),
                     legendgroup=time_frame
                 )
                 d_traces[weather_year][time_frame]['Error'] = go.Bar(
                     x=d_predictions[weather_year][time_frame]['Date'],
                     y=d_predictions[weather_year][time_frame]['Error'],
-                    name=time_frame + ' Error ' + weather_year,
-                    legendgroup=time_frame
+                    name="{0} Error ({1} Weather)".format(time_frame, weather_year),
+                    legendgroup=time_frame,
+                    yaxis='y2'
                 )
 
             except KeyError:
@@ -192,33 +214,34 @@ if __name__ == "__main__":
                 d_traces[weather_year][time_frame]['Prediction'] = go.Scatter(
                     x=d_predictions[weather_year][time_frame]['Date'],
                     y=d_predictions[weather_year][time_frame]['Prediction'],
-                    name=time_frame + ' Prediction ' + weather_year,
+                    name="{0} Prediction ({1} Weather)".format(time_frame, weather_year),
                     line=dict(dash='dash'),
                     legendgroup=time_frame
                 )
                 d_traces[weather_year][time_frame]['Error'] = go.Bar(
                     x=d_predictions[weather_year][time_frame]['Date'],
                     y=d_predictions[weather_year][time_frame]['Error'],
-                    name=time_frame + ' Error ' + weather_year,
-                    legendgroup=time_frame
+                    name="{0} Error ({1} Weather)".format(time_frame, weather_year),
+                    legendgroup=time_frame,
+                    yaxis='y2'
                 )
 
     # Create traces for actual data
     trace_yearly = go.Scatter(
-        x=d_predictions['2016']['Yearly']['Date'],
-        y=d_predictions['2016']['Yearly']['ACT'],
+        x=d_predictions['Actual']['Yearly']['Date'],
+        y=d_predictions['Actual']['Yearly']['ACT'],
         name='Yearly Actual',
         legendgroup='Yearly',
     )
     trace_monthly = go.Scatter(
-        x=d_predictions['2016']['Monthly']['Date'],
-        y=d_predictions['2016']['Monthly']['ACT'],
+        x=d_predictions['Actual']['Monthly']['Date'],
+        y=d_predictions['Actual']['Monthly']['ACT'],
         name='Monthly Actual',
         legendgroup='Monthly',
     )
     trace_daily = go.Scatter(
-        x=d_predictions['2016']['Daily']['Date'],
-        y=d_predictions['2016']['Daily']['ACT'],
+        x=d_predictions['Actual']['Daily']['Date'],
+        y=d_predictions['Actual']['Daily']['ACT'],
         name='Daily Actual',
         legendgroup='Daily',
     )
@@ -246,8 +269,9 @@ if __name__ == "__main__":
     # Create layout for plotly
     layout = go.Layout(
         title='URD Prediction vs. Actual',
-        xaxis=dict(title='', rangeslider=dict(thickness=0.1), type='date', showgrid=True),
-        yaxis=dict(title='', showgrid=True),
+        xaxis=dict(title='', rangeslider=dict(thickness=0.015, borderwidth=1), type='date', showgrid=True),
+        yaxis=dict(title='', showgrid=True, domain=[0.35, 1]),
+        yaxis2=dict(domain=[0, 0.25]),
         updatemenus=list([
             dict(
                 buttons=[vis_dict for vis_dict in l_vis_dicts],
